Fix bug in parsing of comments.
When parsing a line comment `// ...` that ends without a newline (because it
ends with the input), the parser was neglecting to ensure that there are no
unsafe characters in the comment. This commit adds that check, and also adds a
test that would fail without this fix.

Also improve and fix some documentation comments in the code, both in the file
that was fixed and in a couple of other files.