[Fix] Add deserialization check on max solutions.
During deserialization, it is important to check that the number of next items
to deserialize is within bounds. For block deserialization, this is done for
transactions, aborted transaction IDs, and and aborted solution IDs, but the
check for solutions was missing. This commit adds it.

Also fix a comment in some deserialization code.