Introduce the dev_println and dev_eprintln macros.
If the `dev-print` feature is enabled, these work like
`println` and `eprintln`, respectively. This is to give
users (in particular Leo) more control of snarkvm's output -
currently some print statements are compiled in when
`debug-assertions` is on, but we want to elide those print statements
even with `debug-assertions`.

I also removed some superfluous `#[must_use]` attributes - this
attribute does nothing when used on a function in a trait
implementation.